+++ libft/sources/strings/ft_strncmp.c
@@ -6,13 +6,13 @@
 /*   By: vtestut <vtestut@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/05 15:51:33 by vtestut           #+#    #+#             */
-/*   Updated: 2024/01/22 13:22:53 by vtestut          ###   ########.fr       */
+/*   Updated: 2024/02/11 15:15:52 by vtestut          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_string.h"
 
-int	ft_strncmp(const char *s1, const char *s2, int n)
+int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
 	if (s1 == NULL || s2 == NULL)
 		return (0);
