+++ libft/sources/strings/ft_substr.c
@@ -6,7 +6,7 @@
 /*   By: vtestut <vtestut@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/01 20:53:28 by vtestut           #+#    #+#             */
-/*   Updated: 2024/01/30 11:50:51 by vtestut          ###   ########.fr       */
+/*   Updated: 2024/02/11 15:20:14 by vtestut          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,14 +16,14 @@ char	*ft_substr(char const *s, unsigned int start, int len)
 {
 	char	*res;
 	char	*src;
-	int		reslen;
+	size_t	reslen;
 
 	if (!s)
 		return (NULL);
-	if (ft_strlen(s) < (int)start)
+	if (ft_strlen(s) < (size_t)start)
 		return (ft_strdup(""));
 	src = (char *)s + start;
-	if (ft_strlen(src) < len)
+	if (ft_strlen(src) < (size_t)len)
 		reslen = ft_strlen(src) + 1;
 	else
 		reslen = len + 1;
